Revert "Feat: Add proxy to Assets R2 on worker (#412)"
This reverts commit 0f39c07e9032878d071816c61c31e351ca96fdf2.
