+++ apps/api/src/app.ts
@@ -2,7 +2,6 @@ import { Hono } from "hono";
 import { timing } from "hono/timing";
 import api from "./api.ts";
 import apps from "./apps.ts";
-import assets from "./assets.ts";
 import { AppEnv } from "./utils/context.ts";
 import { Hosts } from "@deco/sdk/hosts";
 
@@ -24,7 +23,6 @@ const normalizeHost = (req: Request) => {
   }
   return {
     [Hosts.API]: Hosts.API,
-    [Hosts.ASSETS]: Hosts.ASSETS,
     localhost: Hosts.API,
     "localhost:3001": Hosts.API,
     "localhost:8000": Hosts.API,
@@ -42,6 +40,5 @@ app.use(timing({ crossOrigin: true, total: true }));
 
 app.route(`/${Hosts.API}`, api);
 app.route(`/${Hosts.APPS}`, apps);
-app.route(`/${Hosts.ASSETS}`, assets);
 
 export default app;
