+++ packages/sdk/src/constants.ts
@@ -35,8 +35,6 @@ export const AUTH_URL = LOCAL_DEBUGGER
   ? "http://localhost:3001"
   : "https://api.deco.chat";
 
-export const ASSETS_URL = "https://assets.deco.chat";
-
 export const API_HEADERS = {
   "content-type": "application/json",
   "use-api-host": "true",
@@ -46,8 +44,6 @@ export const WELL_KNOWN_AGENT_IDS = {
   teamAgent: "teamAgent",
   setupAgent: "setupAgent",
 } as const;
-
-export const PUBLIC_ASSETS_BUCKET = "public";
 
 export interface Model {
   id: string;
