+++ packages/sdk/src/hosts.ts
@@ -1,7 +1,6 @@
 export const Hosts = {
   API: "api.deco.chat",
   APPS: "deco.page",
-  ASSETS: "assets.deco.chat",
   FS: "fs.deco.chat",
   Chat: "deco.chat",
   LOCALHOST: "localhost:3000",
