+++ packages/sdk/src/mcp/fs/api.ts
@@ -3,16 +3,75 @@ import {
   GetObjectCommand,
   ListObjectsCommand,
   PutObjectCommand,
+  S3Client,
 } from "@aws-sdk/client-s3";
 import { getSignedUrl } from "@aws-sdk/s3-request-presigner";
 import { z } from "zod";
-import { getWorkspaceBucketName } from "../../storage/s3/utils.ts";
+import { WELL_KNOWN_ORIGINS } from "../../hosts.ts";
 import {
   assertHasWorkspace,
   canAccessWorkspaceResource,
 } from "../assertions.ts";
-import { createTool } from "../context.ts";
-import { ensureBucketExists } from "../utils.ts";
+import { AppContext, createTool, getEnv } from "../context.ts";
+
+// Regex generated by Cursor on 2025-05-21
+// Replace any character that is not a letter, number, or hyphen with a hyphen to not break the bucket name
+const getWorkspaceBucketName = (c: AppContext) =>
+  `deco-chat-${
+    c.workspace!.value
+      .toLowerCase()
+      .replace(/[^a-z0-9]/g, "-")
+      .replace(/-+/g, "-")
+      .replace(/^-|-$/g, "")
+  }`;
+
+const ensureBucketExists = async (c: AppContext, bucketName: string) => {
+  const { cf } = c;
+  const env = getEnv(c);
+
+  try {
+    await cf.r2.buckets.get(bucketName, {
+      account_id: env.CF_ACCOUNT_ID,
+    });
+  } catch (error) {
+    if ((error as unknown as { status: number })?.status !== 404) {
+      throw error;
+    }
+
+    // Create bucket
+    await cf.r2.buckets.create({
+      name: bucketName,
+      account_id: env.CF_ACCOUNT_ID,
+    });
+
+    // Set cors
+    await cf.r2.buckets.cors.update(bucketName, {
+      account_id: env.CF_ACCOUNT_ID,
+      rules: [{
+        maxAgeSeconds: 3600,
+        exposeHeaders: ["etag"],
+        allowed: {
+          methods: ["GET", "PUT"],
+          origins: [...WELL_KNOWN_ORIGINS],
+          headers: ["origin", "content-type"],
+        },
+      }],
+    });
+  }
+};
+
+const getS3Client = (c: AppContext) => {
+  const env = getEnv(c);
+
+  return new S3Client({
+    region: "auto",
+    endpoint: `https://${env.CF_ACCOUNT_ID}.r2.cloudflarestorage.com`,
+    credentials: {
+      accessKeyId: env.CF_R2_ACCESS_KEY_ID!,
+      secretAccessKey: env.CF_R2_SECRET_ACCESS_KEY!,
+    },
+  });
+};
 
 export const listFiles = createTool({
   name: "FS_LIST",
@@ -22,17 +81,18 @@ export const listFiles = createTool({
   }),
   canAccess: canAccessWorkspaceResource,
   handler: async ({ prefix: root }, c) => {
-    assertHasWorkspace(c);
-    const bucketName = getWorkspaceBucketName(c.workspace.value);
+    const bucketName = getWorkspaceBucketName(c);
 
+    assertHasWorkspace(c);
     await ensureBucketExists(c, bucketName);
 
+    const s3Client = getS3Client(c);
     const listCommand = new ListObjectsCommand({
       Bucket: bucketName,
       Prefix: root,
     });
 
-    return c.s3.send(listCommand);
+    return s3Client.send(listCommand);
   },
 });
 
@@ -47,17 +107,18 @@ export const readFile = createTool({
   }),
   canAccess: canAccessWorkspaceResource,
   handler: async ({ path, expiresIn = 60 }, c) => {
-    assertHasWorkspace(c);
-    const bucketName = getWorkspaceBucketName(c.workspace.value);
+    const bucketName = getWorkspaceBucketName(c);
 
+    assertHasWorkspace(c);
     await ensureBucketExists(c, bucketName);
 
+    const s3Client = getS3Client(c);
     const getCommand = new GetObjectCommand({
       Bucket: bucketName,
       Key: path,
     });
 
-    const url = await getSignedUrl(c.s3, getCommand, { expiresIn });
+    const url = await getSignedUrl(s3Client, getCommand, { expiresIn });
 
     return { url };
   },
@@ -71,17 +132,18 @@ export const readFileMetadata = createTool({
   }),
   canAccess: canAccessWorkspaceResource,
   handler: async ({ path }, c) => {
-    assertHasWorkspace(c);
-    const bucketName = getWorkspaceBucketName(c.workspace.value);
+    const bucketName = getWorkspaceBucketName(c);
 
+    assertHasWorkspace(c);
     await ensureBucketExists(c, bucketName);
 
+    const s3Client = getS3Client(c);
     const getCommand = new GetObjectCommand({
       Bucket: bucketName,
       Key: path,
     });
 
-    const response = await c.s3.send(getCommand);
+    const response = await s3Client.send(getCommand);
 
     return {
       metadata: response.Metadata,
@@ -106,11 +168,12 @@ export const writeFile = createTool({
   }),
   canAccess: canAccessWorkspaceResource,
   handler: async ({ path, expiresIn = 60, contentType, metadata }, c) => {
-    assertHasWorkspace(c);
-    const bucketName = getWorkspaceBucketName(c.workspace.value);
+    const bucketName = getWorkspaceBucketName(c);
 
+    assertHasWorkspace(c);
     await ensureBucketExists(c, bucketName);
 
+    const s3Client = getS3Client(c);
     const putCommand = new PutObjectCommand({
       Bucket: bucketName,
       Key: path,
@@ -118,7 +181,7 @@ export const writeFile = createTool({
       Metadata: metadata,
     });
 
-    const url = await getSignedUrl(c.s3, putCommand, {
+    const url = await getSignedUrl(s3Client, putCommand, {
       expiresIn,
       signableHeaders: new Set(["content-type"]),
     });
@@ -133,15 +196,17 @@ export const deleteFile = createTool({
   inputSchema: z.object({ path: z.string() }),
   canAccess: canAccessWorkspaceResource,
   handler: async ({ path }, c) => {
-    assertHasWorkspace(c);
-    const bucketName = getWorkspaceBucketName(c.workspace.value);
+    const bucketName = getWorkspaceBucketName(c);
 
+    assertHasWorkspace(c);
     await ensureBucketExists(c, bucketName);
 
+    const s3Client = getS3Client(c);
     const deleteCommand = new DeleteObjectCommand({
       Bucket: bucketName,
       Key: path,
     });
-    return c.s3.send(deleteCommand);
+
+    return s3Client.send(deleteCommand);
   },
 });
