+++ packages/sdk/src/storage/index.ts
@@ -1,4 +1,2 @@
 export * from "./supabase/schema.ts";
 export * from "./supabase/client.ts";
-export * from "./s3/client.ts";
-export * from "./s3/utils.ts";
