--- packages/sdk/src/storage/s3/utils.ts
@@ -1,14 +1,0 @@
-/**
- * Generates a bucket name for a workspace
- * @param workspaceValue The workspace value
- * @returns Normalized bucket name
- */
-export function getWorkspaceBucketName(workspaceValue: string): string {
-  return `deco-chat-${
-    workspaceValue
-      .toLowerCase()
-      .replace(/[^a-z0-9]/g, "-")
-      .replace(/-+/g, "-")
-      .replace(/^-|-$/g, "")
-  }`;
-}
